% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Npres_Fucntions.R
\name{Calc.F.cond}
\alias{Calc.F.cond}
\title{Conditional Distribution calculation at the data points}
\usage{
Calc.F.cond(x, y, z, z.dim, bwmet, bwobj = NULL)
}
\arguments{
\item{x}{A numeric vector of observations of x}

\item{y}{A numeric vector of observations of y}

\item{z}{A numeric matrix of observations of z}

\item{z.dim}{Dimension of z, number of columns ofz matrix}

\item{bwmet}{which method to use to select bandwidths. cv.ls specifies least-squares cross-validation, and normal-reference just computes the ‘rule-of-thumb’ bandwidth hj using the standard formula hj = 1.06σj n−1/(2P +l) , where σj is an adaptive measure of spread of the jth continuous variable defined as min(standard deviation, mean absolute de- viation/1.4826, interquartile range/1.349), n the number of observations, P the order of the kernel, and l the number of continuous variables. Note that when there exist factors and the normal-reference rule is used, there is zero smoothing of the factors. Defaults to cv.ml.}

\item{bwobj}{should be of the kind that is returned from  Calc.F.cond. Use it in case you want to use previously calculated bandwdiths  (from other data points) to evaluate the estimate the conditional distribution functions from (x,y,z). Default is NULL. In that case uses `bwmet` to compute this#'}
}
\value{
A list including the elements:
\item{alp.hat}{estimate of alpha. NULL if `method' is `lwr.bnd'.}
\item{alp.Lwr}{95\% lower confidence bound of alpha. NULL if `method` is `cv` or `fixed`.}
\item{Fs.hat}{A list containing with elements `x` and `y` values for the function estimate of `F.s`}
\item{dist.out}{An object of the class `dist.calc` using the complete data.gen}
\item{c.n}{ Value of the tuning parameter used to compute the final estimate.}
\item{cv.out}{An object of class `cv.mix.model`. The object is NULL if method is `fixed`.}

\item{F.x_z}{Estimate of conditional distribution of X|Z}
\item{F.y_z}{Estimate of conditional distribution of Y|Z}
\item{Fbw.x_z}{Bandwidths used for computation of conditional distribution of x|z}
\item{Fbw.y_z}{Bandwidths used for computation of conditional distribution of y|z}
\item{Fbw.z_z}{Bandwidths used for computation of conditional distribution of z|z}
\item{F.z_hat}{Estimate of conditional distribution of Z_d|Z_{1,..,(d-1)}, Z_{d-1}|Z_{1..(d-2)},..., Z_2|Z_1,}
}
\description{
Conditional Distribution calculation at the data points
}
